+++ packaging/chocolatey/tools/chocolateyInstall.ps1
@@ -1,11 +1,9 @@
 $ErrorActionPreference = 'Stop'
 
 $packageName  = 'localpdf-studio'
-$version      = '0.10.0'
+$version      = '0.15.0'
 $url64        = "https://github.com/jeiel85/localpdf-studio/releases/download/v$version/LocalPDF.Studio_${version}_x64-setup.exe"
-# NOTE: PR 제출 시 release 게시 후 실제 SHA256으로 교체.
-#   $sha = (Get-FileHash .\LocalPDF.Studio_$version_x64-setup.exe -Algorithm SHA256).Hash
-$checksum64   = 'REPLACE_WITH_ACTUAL_SHA256_FROM_RELEASE_ASSET'
+$checksum64   = 'c76076008e979ce653d864b6549e7f813c3e1bfd40fa1320019671e3195fa6a5'
 $checksumType = 'sha256'
 
 $packageArgs = @{